Add Swift Package Manager Support
